requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.5
contourpy==1.0.5
cryptography==37.0.1
cycler==0.11.0
Cython==0.29.32
cytoolz==0.11.0
dask==2022.7.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
fonttools==4.38.0
frozenlist==1.2.0
fsspec==2022.7.1
future==0.18.2
google-auth==2.6.0
google-auth-oauthlib==0.4.4
grpcio==1.42.0
idna==3.4
imagecodecs==2021.8.26
imageio==2.19.3
importlib-metadata==4.11.3
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.16.0
jupyter_client==7.3.5
jupyter_core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.1.2
jupyterlab_server==2.15.2
kiwisolver==1.4.4
locket==1.0.0
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
nbclassic==0.3.5
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==2.8.4
notebook==6.4.12
numpy==1.21.0
oauthlib==3.2.1
ortools==9.4.1874
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
prometheus-client==0.14.1
prompt-toolkit==3.0.20
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
pywin32==302
pywinpty==2.0.2
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.19.2
scipy==1.9.1
Send2Trash==1.8.0
setuptools==57.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tensorboard==2.9.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
terminado==0.13.1
testpath==0.6.0
tifffile==2021.7.2
toolz==0.11.2
torch==1.10.2
torchvision==0.11.0a0
tornado==6.2
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.0.3
wheel==0.36.2
win-inet-pton==1.1.0
wincertstore==0.2
yarl==1.8.1
zipp==3.8.0
